Add files via upload
This Python program uses a while loop to collect scores and calculates the scores average after dropping the lowest grade.  The scores and validated and errored out if they are not valid and prints output of the lowest score, the modified list of scores, score average and the letter grade.