Add files via upload
This Python program uses a while loop and if else statements to calculate employee salary and overtime and prints  the Hours worked, Pay Rate, Overtime, Overtime Pay, Regular Hour pay and the  Gross Pay  and displays in a one line format.  It calculates the Total number of employees entered, Total amount paid for overtime, Total amount paid for regular hours, Total amount paid in gross and then prints a display at the end of the program.