Add files via upload
This program that uses loops, functions and module import  to create  simple math quizzes.  The program displays two random numbers and then the program allows the user to enter the answer.   If the answer is correct, a message of congratulations should be displayed.  If the answer is incorrect, the program notifies the user if their answer was too low or too high and ask them to guess again. The user will have to continue to guess until they get the right answer.  The program keeps track of the number of guesses.  Once the user has made a correct guess, the program is tells them the number of guesses it took to get the right answer.