Add files via upload
This program creates a Math Quiz and selects two random numbers to add or subtract and asks the user to give the correct answer.  A menu is displayed that asks the user if they would like to add, subtract or exit and the user makes a selection.   If the user gives an answer that is too low the program alerts the user the answer is too low and to try again.  If the answer is too high the program alerts the user the answer is too high and to try again.  If the user gets the correct answer the program congratulates the user on getting the correct answer and tells the user how many guesses it took to get the correct answer.  The program ends if the user selects exit.